refactor(server): replace custom MinioSupervisor with minio_server package
- Remove custom MinioSupervisor implementation
- Add minio_server dependency (~> 0.1.0) for dev environment
- Configure MinioServer with extracted port from s3_endpoint
- Add MinioBucketCreator task to create default bucket after startup
- Simplifies MinIO lifecycle management using established library
